Feature: Redesign Active Loans table using SHADCN and TailwindCSS
